Creating server module and API docs stub
