+++ docs/developers/api/server.md
@@ -1,0 +1,11 @@
+# The `server` submodule
+
+::: safedata_validator.server
+    options:
+        show_source: true
+        group_by_category: false
+        members:
+            - post_metadata
+            - update_gazetteer
+        show_root_heading: false
+        show_root_toc_entry: false
