+++ safedata_validator/zenodo.py
@@ -536,78 +536,6 @@ def delete_file(
         return {"result": "success"}, None
 
 
-# SERVER COMMANDS
-
-
-def post_metadata(
-    metadata: dict, zenodo: dict, resources: Optional[Resources] = None
-) -> ZenodoFunctionResponseType:
-    """Post the dataset metadata and zenodo metadata to the metadata server.
-
-    Args:
-        metadata: The dataset metadata dictionary for a dataset
-        zenodo: The dataset metadata dictionary for a deposit
-        resources: The safedata_validator resource configuration to be used. If
-            none is provided, the standard locations are checked.
-
-    Returns:
-        See [here][safedata_validator.zenodo.ZenodoFunctionResponseType].
-    """
-
-    # Get resource configuration
-    zres = _resources_to_zenodo_api(resources)
-
-    payload = {"metadata": metadata, "zenodo": zenodo}
-
-    # post the metadata to the server
-    mtd = requests.post(
-        f"{zres['mdapi']}/post_metadata",
-        params={"token": zres["mdtoken"]},
-        json=payload,
-        verify=zres["mdssl"],
-    )
-
-    # trap errors in uploading metadata and tidy up
-    if mtd.status_code != 201:
-        return {}, mtd.text
-    else:
-        return mtd.json(), None
-
-
-def update_gazetteer(
-    gazetteer: dict, location_aliases: dict, resources: Optional[Resources] = None
-) -> ZenodoFunctionResponseType:
-    """Update the gazetteer and location aliases used by the metadata server.
-
-    Args:
-        gazetteer: Updated gazetteer data in GeoJSON format.
-        location_aliases: Updated location aliases in JSON format.
-        resources: The safedata_validator resource configuration to be used. If
-            none is provided, the standard locations are checked.
-
-    Returns:
-        See [here][safedata_validator.zenodo.ZenodoFunctionResponseType].
-    """
-
-    # Get resource configuration
-    zres = _resources_to_zenodo_api(resources)
-
-    payload = {"gazetteer": gazetteer, "location_aliases": location_aliases}
-
-    # post the metadata to the server
-    response = requests.post(
-        f"{zres['mdapi']}/update_gazetteer",
-        params={"token": zres["mdtoken"]},
-        json=payload,
-    )
-
-    # trap errors in uploading metadata and tidy up
-    if response.status_code != 201:
-        return {}, response.text
-    else:
-        return response.json(), None
-
-
 """
 Dataset description generation (HTML and GEMINI XML)
 """
